+++ backend/src/main/java/de/iav/backend/repository/AppUserRepository.java
@@ -1,6 +1,6 @@
 package de.iav.backend.repository;
 
-import de.iav.backend.model.AppUser;
+import de.iav.backend.security.AppUser;
 import de.iav.backend.model.Ticket;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
